Rename audio::hacks module to audio::stderr_silencer
