+++ src/audio/stderr_silencer.rs
